+++ Sandboxie/apps/start/aboutdlg.cpp
@@ -307,21 +307,11 @@ bool DoAboutDialog(bool bReminder)
             if (Days < 40)
                 return true;
 
-            int Probability = 5;
-			int Interval; // days
-            if (Days > 369) {
-                Interval = 3;
-            //    Probability = 10;
-            } else if (Days > 120) {
-                Interval = 5;
-            //    Probability = 20;
-            } else if (Days > 80) {
-                Interval = 7;
-            //    Probability = 40;
-            } else {//if (Days > 40)	
-                Interval = 10;
-            //    Probability = 80;
-            }
+		    int Interval; // days
+            if (Days > 730) Interval = 5;
+            else if (Days > 365) Interval = 10;
+            else if (Days > 180) Interval = 20;
+            else Interval = 30;
 
 			time_t LastReminder = 0;
             SbieApi_Call(API_GET_SECURE_PARAM, 3, L"LastReminder", (ULONG_PTR)&LastReminder, sizeof(LastReminder));
@@ -330,7 +320,7 @@ bool DoAboutDialog(bool bReminder)
 					return true;
 			}
 
-            if ((rand() % Probability) != 0)
+            if ((rand() % 5) != 0)
                 return true;
         }