+++ SandboxiePlus/SandMan/OnlineUpdater.h
@@ -5,6 +5,23 @@
 
 #include "SbiePlusAPI.h"
 
+class CGetUpdatesJob : public QObject
+{
+	Q_OBJECT
+
+protected:
+	friend class COnlineUpdater;
+
+	CGetUpdatesJob(const QVariantMap& Params, QObject* parent = nullptr) : QObject(parent) { m_Params = Params; }
+	virtual ~CGetUpdatesJob() {}
+
+	QVariantMap	m_Params;
+
+signals:
+	void		UpdateData(const QVariantMap& Data, const QVariantMap& Params);
+};
+
+
 class COnlineUpdater : public QObject
 {
 	Q_OBJECT
@@ -15,6 +32,8 @@ public:
 
 	static void			Process();
 
+	void				GetUpdates(QObject* receiver, const char* member, const QVariantMap& Params = QVariantMap());
+
 	void				InstallUpdate();
 
 	void				UpdateCert();
@@ -23,6 +42,8 @@ public:
 	void				DownloadUpdates(const QString& DownloadUrl, bool bManual);
 
 private slots:
+	
+	void				OnUpdateData(const QVariantMap& Data, const QVariantMap& Params);
 
 	void				OnUpdateCheck();
 	void				OnUpdateProgress(qint64 bytes, qint64 bytesTotal);
@@ -34,4 +55,5 @@ protected:
 
 	CNetworkAccessManager*	m_RequestManager;
 	CSbieProgressPtr	m_pUpdateProgress;
+	QMap<QNetworkReply*, CGetUpdatesJob*> m_JobQueue;
 };